Issue #41 - Fix skins installing into the wrong directory
The vendor directory into which extensions get installed is set once and sub-
sequently cached in the ExtensionInstaller class, from which both the Plugin-
and the SkinInstaller classes inherit.

This code used to work as intended, however PHP 8.1 changed the behaviour [1]
and now all child classes share the static variable they inherited from their
parent class. This means that when the plugin installer needs to install both
regular plugins *and* skins within the same install session, the same path ends
up being used both for skins and plugins, so one or the other ends up in the
wrong directory.

The fix for this is based on the proposed workaround contained in the respective
PHP RFC at [1].

[1] https://wiki.php.net/rfc/static_variable_inheritance